requirements.txt
numpy
pandas
streamlit
pillow
scikit-learn==1.3.0
nltk==3.8.1